backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
alembic==1.13.1

pydantic==1.10.13
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
email-validator==2.1.1

tensorflow-cpu==2.15.0
pillow==10.2.0
numpy==1.26.4
